moved SyntaxBuilder to Syntax namespace with maintained BC
