+++ phpunit.xml.dist
@@ -19,9 +19,11 @@
     </testsuites>
 
     <logging>
+        <!--
         <log type="coverage-html" target="coverage" charset="UTF-8"
              yui="true" highlight="false" lowUpperBound="50" highLowerBound="90"/>
         <log type="coverage-clover" target="coverage.xml"/>
+        -->
     </logging>
 
     <filter>
