+++ src/SyntaxBuilder.php
@@ -4,64 +4,65 @@ namespace Thunder\Shortcode;
 /**
  * @author Tomasz Kowalczyk <tomasz@kowalczyk.cc>
  */
-final class SyntaxBuilder
+class_alias('Thunder\\Shortcode\\Syntax\\SyntaxBuilder', 'Thunder\\Shortcode\\SyntaxBuilder', true);
+return;
+
+/**
+ * Weird, PHP does not see the aliased class and does not honor return statement,
+ * SyntaxBuilder declaration needs to be wrapped as shown below to work...
+ */
+if(false)
     {
-    private $openingTag;
-    private $closingTag;
-    private $closingTagMarker;
-    private $parameterValueSeparator;
-    private $parameterValueDelimiter;
-
-    public function __construct()
-        {
-        }
-
-    public function getSyntax()
-        {
-        return new Syntax($this->openingTag, $this->closingTag, $this->closingTagMarker,
-            $this->parameterValueSeparator, $this->parameterValueDelimiter);
-        }
-
     /**
-     * @deprecated
+     * This implementation is left only to not break IDE autocompletion, this class
+     * is deprecated, it was moved to the new location as specified in docblock.
+     * This file will be removed in version 1.0!
+     *
+     * @deprecated use Thunder\Shortcode\Syntax\SyntaxBuilder
+     * @codeCoverageIgnore
      */
-    public function setStrict($isStrict)
+    class SyntaxBuilder
         {
-        return $this;
-        }
+        public function __construct()
+            {
+            }
 
-    public function setOpeningTag($tag)
-        {
-        $this->openingTag = $tag;
+        public function getSyntax()
+            {
+            return new Syntax();
+            }
 
-        return $this;
-        }
+        /**
+         * @deprecated
+         */
+        public function setStrict($isStrict)
+            {
+            return $this;
+            }
 
-    public function setClosingTag($tag)
-        {
-        $this->closingTag = $tag;
+        public function setOpeningTag($tag)
+            {
+            return $this;
+            }
 
-        return $this;
-        }
+        public function setClosingTag($tag)
+            {
+            return $this;
+            }
 
-    public function setClosingTagMarker($marker)
-        {
-        $this->closingTagMarker = $marker;
+        public function setClosingTagMarker($marker)
+            {
+            return $this;
+            }
 
-        return $this;
-        }
+        public function setParameterValueSeparator($separator)
+            {
+            return $this;
+            }
 
-    public function setParameterValueSeparator($separator)
-        {
-        $this->parameterValueSeparator = $separator;
-
-        return $this;
-        }
-
-    public function setParameterValueDelimiter($delimiter)
-        {
-        $this->parameterValueDelimiter = $delimiter;
-
-        return $this;
+        public function setParameterValueDelimiter($delimiter)
+            {
+            return $this;
+            }
         }
     }
